+++ ThankingHack/Utilities/SphereUtilities.cs
@@ -18,7 +18,7 @@ namespace Thanking.Utilities
 			List<Vector3> nVerts = new List<Vector3>();
 
 			for (int i = 0; i < verts.Length; i++)
-				if (!Physics.Raycast(pos, (go.transform.TransformPoint(verts[i]) - pos).normalized, VectorUtilities.GetDistance(pos, go.transform.TransformPoint(verts[i])) + 0.5f, mask))
+				if (!Physics.Raycast(pos, (go.transform.TransformPoint(verts[i]) - pos).normalized, (float)(VectorUtilities.GetDistance(pos, go.transform.TransformPoint(verts[i])) + 0.5), mask))
 					return go.transform.TransformPoint(verts[i]);
 			
             return Vector3.zero;
